Fix reviews view and serializers to aid in PUT/DELETE requests

+++ reviews/views.py
@@ -1,34 +1,51 @@
-# reviews/views.py
-from rest_framework import status, generics, filters
-from rest_framework.views import APIView
-from rest_framework.response import Response
+from rest_framework.generics import ListAPIView, RetrieveUpdateDestroyAPIView, CreateAPIView
 from rest_framework.permissions import IsAuthenticated
+from rest_framework.response import Response
+from rest_framework import status
+from django.http import Http404
 from .models import Review
-from books.models import Book
 from .serializers import ReviewSerializer
-from main.permissions import IsAuthorOrReadOnly, IsOwnerOrReadOnly
+from main.permissions import IsAuthorOrReadOnly
 
 
-from django.db.models import Count
-
-
-class ReviewList(generics.ListAPIView):
-    queryset = Review.objects.annotate(likes_count=Count('like')).all()
+class ProfileReviews(ListAPIView):
     serializer_class = ReviewSerializer
-    filter_backends = [
-        filters.OrderingFilter,
-        filters.SearchFilter,
-    ]
-    search_fields = [
-        'owner__username',
-        'book__title'
-    ]
-    ordering_fields = [
-        'likes_count'
-    ]
+    permission_classes = [IsAuthenticated]
+
+    def get_queryset(self):
+        owner_id = self.kwargs.get('owner_id')
+        if owner_id is not None:
+            return Review.objects.filter(owner__id=owner_id)
+        return Review.objects.none()
 
 
-class ReviewDetail(APIView):
+class BookReviews(ListAPIView):
+    serializer_class = ReviewSerializer
+    permission_classes = [IsAuthenticated]
+
+    def get_queryset(self):
+        pk = self.kwargs.get('pk')
+        if pk is not None:
+            return Review.objects.filter(book__id=pk)
+        return Review.objects.none()
+
+
+class ReviewCreate(CreateAPIView):
+    queryset = Review.objects.all()
+    serializer_class = ReviewSerializer
+    permission_classes = [IsAuthenticated]
+
+    def perform_create(self, serializer):
+        serializer.save(owner=self.request.user)
+
+
+class ReviewList(ListAPIView):
+    queryset = Review.objects.all()
+    serializer_class = ReviewSerializer
+
+
+class ReviewDetail(RetrieveUpdateDestroyAPIView):
+    queryset = Review.objects.all()
     serializer_class = ReviewSerializer
     permission_classes = [IsAuthorOrReadOnly]
 
@@ -40,14 +57,12 @@ class ReviewDetail(APIView):
 
     def get(self, request, pk):
         review = self.get_object(pk)
-        review_with_likes_count = Review.objects.annotate(
-            likes_count=Count('like')).get(pk=pk)
-        serializer = ReviewSerializer(
-            review_with_likes_count, context={'request': request})
+        serializer = ReviewSerializer(review)
         return Response(serializer.data)
 
     def put(self, request, pk):
         review = self.get_object(pk)
+        self.check_object_permissions(request, review)
         serializer = ReviewSerializer(review, data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -56,49 +71,6 @@ class ReviewDetail(APIView):
 
     def delete(self, request, pk):
         review = self.get_object(pk)
+        self.check_object_permissions(request, review)
         review.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
-
-
-class ProfileReviews(APIView):
-    permission_classes = [IsAuthorOrReadOnly]
-    serializer_class = ReviewSerializer
-
-    def get(self, request, owner_id):
-        reviews = Review.objects.filter(owner_id=owner_id)
-        serializer = ReviewSerializer(reviews, many=True)
-        return Response(serializer.data)
-
-
-class BookReviews(APIView):
-    permission_classes = [IsAuthorOrReadOnly]
-    serializer_class = ReviewSerializer
-
-    def get(self, request, pk):
-        try:
-            book = Book.objects.get(pk=pk)
-            reviews = Review.objects.filter(book=book)
-            serializer = ReviewSerializer(
-                reviews, many=True, context={'request': request})
-            return Response(serializer.data)
-        except Book.DoesNotExist:
-            return Response(status=404)
-
-
-class ReviewCreate(generics.CreateAPIView):
-    queryset = Review.objects.all()
-    serializer_class = ReviewSerializer
-    permission_classes = [IsAuthenticated]
-
-    def perform_create(self, serializer):
-        serializer.save(owner=self.request.user)
-
-
-class ProfileReviews(generics.ListAPIView):
-    serializer_class = ReviewSerializer
-
-    def get_queryset(self):
-        owner_id = self.kwargs.get('owner_id')
-        if owner_id is not None:
-            return Review.objects.filter(owner__id=owner_id).annotate(likes_count=Count('like'))
-        return Review.objects.none()
